refactor(knowledge): 重构知识库向量搜索功能
-移除 EmbeddingAddCmdExe 类
- 更新 EmbeddingDTO 类，将 vector 字段类型改为 float[]
- 修改 EmbeddingDO 类，将 vector 字段类型改为 float[] 并添加 typeHandler
-重构 EmbeddingGatewayImpl 类，支持 float[] 向量数据
- 新增 JedisConfig 类配置 Redis 连接
- 新增 JedisTextSearchGatewayImpl 类实现向量搜索
- 更新相关查询和命令执行器以支持新的向量数据结构

+++ smartcs-web-app/src/main/java/com/leyue/smartcs/knowledge/listener/DocEmbeddingGenerateListener.java
@@ -154,7 +154,7 @@ public class DocEmbeddingGenerateListener {
         
         try {
             // 调用LLM服务生成向量
-            List<byte[]> vectors = llmGateway.generateEmbeddings(contentChunks);
+            List<float[]> vectors = llmGateway.generateEmbeddings(contentChunks);
             
             // 构建Embedding对象列表
             for (int i = 0; i < contentChunks.size(); i++) {
